+++ rdmo/tasks/tests/test_views.py
@@ -1,0 +1,58 @@
+from django.test import TestCase
+
+from test_generator.views import TestListViewMixin
+
+from rdmo.core.testing.mixins import TestExportViewMixin, TestImportViewMixin
+from rdmo.accounts.utils import set_group_permissions
+
+
+class TasksViewTestCase(TestCase):
+
+    fixtures = (
+        'users.json',
+        'groups.json',
+        'accounts.json',
+        'conditions.json',
+        'domain.json',
+        'options.json',
+        'tasks.json',
+    )
+
+    users = (
+        ('editor', 'editor'),
+        ('reviewer', 'reviewer'),
+        ('user', 'user'),
+        ('api', 'api'),
+        ('anonymous', None),
+    )
+
+    status_map = {
+        'list_view': {
+            'editor': 200, 'reviewer': 200, 'api': 200, 'user': 403, 'anonymous': 302
+        },
+        'export_view': {
+            'editor': 200, 'reviewer': 200, 'api': 200, 'user': 403, 'anonymous': 302
+        }
+    }
+
+    @classmethod
+    def setUpTestData(cls):
+        set_group_permissions()
+
+
+class TasksTests(TestListViewMixin, TasksViewTestCase):
+
+    url_names = {
+        'list_view': 'tasks'
+    }
+
+class TasksExportTests(TestExportViewMixin, TasksViewTestCase):
+
+    url_names = {
+        'export_view': 'tasks_export'
+    }
+
+
+class TasksImportTests(TestImportViewMixin, TasksViewTestCase):
+
+    import_file = 'testing/xml/tasks.xml'
